Fixed issue with MakeSparse when accessing files on a remote share.
Fixed an issue with OpenFileOverlappedExclusive that caused the function
to fail and the file handle to be closed if the filesystem cluster size
could not be determined automatically. This modifies the function to
zero the pFsClusterSize if it can't be determined and allows the caller
to use a default value (ie. 4k).